# rule-agent/swagger.yaml
openapi: 3.0.3
info:
  title: Underwriting Rule Generation API
  description: |
    AI-powered underwriting workflow that processes policy documents and generates Drools rules.

    **Key Features:**
    - Extract text from policy PDFs from S3
    - LLM analyzes documents and generates custom extraction queries
    - Extract structured data using AWS Textract (required)
    - Generate Drools DRL rules automatically
    - Deploy to Drools KIE Server with auto-generated container IDs
    - Upload generated artifacts (JAR, DRL, Excel) to S3

    **Prerequisites:**
    - AWS S3 configured and accessible
    - AWS Textract configured with proper IAM permissions
    - Drools KIE Server running and accessible

    **Multi-Tenant Support:**
    - Separate containers per bank and policy type
    - Format: `{bank_id}-{policy_type}-underwriting-rules`

    **Zero Local Storage:**
    - All files use temporary storage and auto-cleanup
    - Generated rules saved to S3 only
  version: 1.0.0
  contact:
    name: IBM Automation
    url: https://github.com/DecisionsDev
servers:
  - url: http://localhost:9000/rule-agent
    description: Local development server
  - url: http://localhost:9000/rule-agent
    description: Docker environment
...
paths:
  /process_policy_from_s3:
    post:
      tags:
        - Underwriting Workflow
      summary: Process policy document from S3
      description: |
        Process a policy PDF directly from S3 URL through the underwriting workflow.
        No file download - reads directly from S3 into memory.
        LLM analyzes the document and generates custom extraction queries automatically.
      operationId: processPolicyFromS3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - s3_url
              properties:
                s3_url:
                  type: string
                  format: uri
                  description: Full S3 URL to the policy PDF
                  example: https://uw-data-extraction.s3.us-east-1.amazonaws.com/policies/chase_insurance_2025.pdf
                policy_type:
                  type: string
                  default: general
                  description: Type of policy
                  example: insurance
                bank_id:
                  type: string
                  description: Bank/tenant identifier (recommended for multi-tenant deployments)
                  example: chase
            examples:
              chase-insurance:
                summary: Chase Insurance from S3
                value:
                  s3_url: https://uw-data-extraction.s3.us-east-1.amazonaws.com/policies/chase/insurance_2025.pdf
                  policy_type: insurance
                  bank_id: chase
              bofa-loan:
                summary: BofA Loan from S3
                value:
                  s3_url: https://uw-data-extraction.s3.us-east-1.amazonaws.com/policies/bofa/loan_policy.pdf
                  policy_type: loan
                  bank_id: bofa
              no-bank:
                summary: Without Bank ID (backwards compatible)
                value:
                  s3_url: https://uw-data-extraction.s3.us-east-1.amazonaws.com/policies/general_policy.pdf
                  policy_type: insurance
      responses:
        '200':
          description: Workflow completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResult'
        '400':
          description: Bad request (missing s3_url, invalid URL)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /drools_container_status:
    get:
      tags:
        - Drools Management
      summary: Get status of specific container
      description: Retrieve detailed status of a specific KIE Server container
      operationId: getContainerStatus
      parameters:
        - name: container_id
          in: query
          required: true
          schema:
            type: string
          description: Container ID to check
          example: chase-insurance-underwriting-rules
      responses:
        '200':
          description: Container status
          content:
            application/json:
              schema:
                type: object
                properties:
                  container-id:
                    type: string
                  status:
                    type: string
                  release-id:
                    type: object
        '404':
          description: Container not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /test_rules:
    post:
      tags:
        - Rule Testing
      summary: Test deployed Drools rules
      description: |
        Execute Drools rules with sample applicant and policy data to test underwriting decisions.

        **How it works:**
        1. Inserts Applicant and Policy facts into Drools working memory
        2. Fires all rules
        3. Retrieves the Decision object with approval status

        **Common Test Scenarios:**
        - Valid applicant (age 18-65, no health conditions, coverage ≤ $1M): Should approve
        - Too young (age < 18): Should reject with age reason
        - Too old (age > 65): Should reject with age reason
        - Health conditions: Should reject with health reason
        - High coverage (> $1M): Should reject with coverage reason
      operationId: testRules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - container_id
                - applicant
                - policy
              properties:
                container_id:
                  type: string
                  description: Drools container ID to test
                  example: chase-insurance-underwriting-rules
                applicant:
                  type: object
                  description: Applicant information
                  required:
                    - name
                    - age
                    - occupation
                  properties:
                    name:
                      type: string
                      example: John Doe
                    age:
                      type: integer
                      minimum: 0
                      maximum: 120
                      example: 35
                    occupation:
                      type: string
                      example: Engineer
                    healthConditions:
                      type: string
                      nullable: true
                      description: Health conditions (null for healthy applicant)
                      example: null
                policy:
                  type: object
                  description: Policy information
                  required:
                    - policyType
                    - coverageAmount
                    - term
                  properties:
                    policyType:
                      type: string
                      example: Term Life
                    coverageAmount:
                      type: number
                      format: double
                      minimum: 0
                      example: 500000
                    term:
                      type: integer
                      minimum: 0
                      example: 20
            examples:
              valid-applicant:
                summary: Valid Applicant (Should Approve)
                description: Healthy 35-year-old with $500K coverage
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: John Doe
                    age: 35
                    occupation: Engineer
                    healthConditions: null
                  policy:
                    policyType: Term Life
                    coverageAmount: 500000
                    term: 20
              too-young:
                summary: Too Young (Should Reject)
                description: 17-year-old applicant - fails age requirement
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: Jane Smith
                    age: 17
                    occupation: Student
                    healthConditions: null
                  policy:
                    policyType: Term Life
                    coverageAmount: 100000
                    term: 10
              too-old:
                summary: Too Old (Should Reject)
                description: 70-year-old applicant - fails age requirement
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: Bob Senior
                    age: 70
                    occupation: Retired
                    healthConditions: null
                  policy:
                    policyType: Term Life
                    coverageAmount: 250000
                    term: 10
              health-conditions:
                summary: Health Conditions (Should Reject)
                description: Applicant with diabetes - fails health check
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: Alice Johnson
                    age: 45
                    occupation: Teacher
                    healthConditions: Diabetes
                  policy:
                    policyType: Term Life
                    coverageAmount: 300000
                    term: 20
              high-coverage:
                summary: High Coverage (Should Reject)
                description: Coverage over $1M - exceeds limit
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: Rich Person
                    age: 40
                    occupation: CEO
                    healthConditions: null
                  policy:
                    policyType: Whole Life
                    coverageAmount: 2000000
                    term: 30
              edge-case-min-age:
                summary: Edge Case - Minimum Age (Should Approve)
                description: 18-year-old applicant - minimum acceptable age
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: Young Adult
                    age: 18
                    occupation: College Student
                    healthConditions: null
                  policy:
                    policyType: Term Life
                    coverageAmount: 100000
                    term: 20
              edge-case-max-age:
                summary: Edge Case - Just Within Max Age (Should Approve)
                description: 64-year-old applicant - just within acceptable age range
                value:
                  container_id: chase-insurance-underwriting-rules
                  applicant:
                    name: Senior Citizen
                    age: 64
                    occupation: Consultant
                    healthConditions: null
                  policy:
                    policyType: Term Life
                    coverageAmount: 500000
                    term: 10
      responses:
        '200':
          description: Rules executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTestResult'
              examples:
                approved:
                  summary: Approved Application
                  value:
                    status: success
                    container_id: chase-insurance-underwriting-rules
                    decision:
                      approved: true
                      reason: Initial evaluation
                      requiresManualReview: false
                      premiumMultiplier: 1.0
                rejected-age:
                  summary: Rejected - Age
                  value:
                    status: success
                    container_id: chase-insurance-underwriting-rules
                    decision:
                      approved: false
                      reason: Applicant age is outside acceptable range
                      requiresManualReview: false
                      premiumMultiplier: 1.0
                rejected-health:
                  summary: Rejected - Health Conditions
                  value:
                    status: success
                    container_id: chase-insurance-underwriting-rules
                    decision:
                      approved: false
                      reason: Applicant has health conditions
                      requiresManualReview: false
                      premiumMultiplier: 1.0
                rejected-coverage:
                  summary: Rejected - Coverage Too High
                  value:
                    status: success
                    container_id: chase-insurance-underwriting-rules
                    decision:
                      approved: false
                      reason: Policy coverage amount is too high
                      requiresManualReview: false
                      premiumMultiplier: 1.0
        '400':
          description: Bad request (missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Container not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    RuleTestResult:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          description: Execution status
          example: success
        container_id:
          type: string
          description: Container ID that was tested
          example: chase-insurance-underwriting-rules
        decision:
          type: object
          description: Underwriting decision from rules execution
          properties:
            approved:
              type: boolean
              description: Whether the application is approved
              example: true
            reason:
              type: string
              description: Explanation for the decision
              example: Initial evaluation
            requiresManualReview:
              type: boolean
              description: Whether manual review is needed
              example: false
            premiumMultiplier:
              type: number
              format: double
              description: Multiplier for premium calculation (1.0 = standard rate)
              example: 1.0
        full_response:
          type: object
          description: Complete Drools KIE Server response (for debugging)
          properties:
            msg:
              type: string
            result:
              type: object
            type:
              type: string
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: No file uploaded
        status:
          type: string
          description: Error status
          example: failed